refactor: unify event types
With this patch, only one event types in required. Software events
are selected with a predicate (see class Event). This allows one
event to be either a hardware event or a software event, depending
on its arguments.

This will be used to unify Read and Exec in a following patch.